requirements.txt
alabaster==0.7.13
asgiref==3.6.0
Babel==2.12.1
certifi==2022.12.7
charset-normalizer==2.1.1
colorama==0.4.6
Django==4.1.1
django-adminlte-3==0.1.6
django-colorfield==0.7.2
django-jazzmin==2.5.0
docutils==0.17.1
fontawesomefree==6.2.0
gunicorn==20.1.0
idna==3.4
imagesize==1.4.1
Jinja2==3.1.2
MarkupSafe==2.1.1
mypy-extensions==0.4.3
packaging==21.3
pathspec==0.10.1
Pillow==9.2.0
platformdirs==2.5.2
psycopg2-binary==2.9.5
Pygments==2.13.0
pyparsing==3.0.9
pytz==2022.5
requests==2.28.1
snowballstemmer==2.2.0
sorl-thumbnail==12.9.0
Sphinx==5.3.0
sphinx-rtd-theme==1.0.0
sphinxcontrib-applehelp==1.0.2
sphinxcontrib-devhelp==1.0.2
sphinxcontrib-htmlhelp==2.0.0
sphinxcontrib-jsmath==1.0.1
sphinxcontrib-qthelp==1.0.3
sphinxcontrib-serializinghtml==1.1.5
sqlparse==0.4.2
tomli==2.0.1
tzdata==2022.2
unicodecsv==0.14.1
urllib3==1.26.12
